requirements.txt
Flask==2.2.5
Pillow==9.4.0
numpy==1.24.3
scipy==1.10.1
tensorflow==2.13.0
